requirements.txt
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1